Added basic User Registration
Form validation for user registration and log in, and action to log out.
Adds authentication to routes, pages can only be seen when user is logged in.
Adds "old('value')" to forms so previous error submissions show.
Adds "exists" requirements for creating contracts.

+++ routes/web.php
@@ -4,31 +4,40 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CustomerController;
 use App\Http\Controllers\ContractController;
 use App\Http\Controllers\ServiceController;
+use App\Http\Controllers\RegisteredUserController;
+use App\Http\Controllers\SessionController;
 use App\Models\Contact;
 
 Route::get('/', function () {
     return view('welcome');
 });
 
-Route::post('/customer', [CustomerController::class, 'create']);
-Route::delete('/customer/{id}', [CustomerController::class, 'destroy']);
+Route::post('/customer', [CustomerController::class, 'create'])->middleware('auth');
+Route::delete('/customer/{id}', [CustomerController::class, 'destroy'])->middleware('auth');
 
-Route::get('/services', [ServiceController::class, 'index']);
-Route::get('/services/{service}', [ServiceController::class, 'show']);
+Route::get('/services', [ServiceController::class, 'index'])->middleware('auth');
+Route::get('/services/{service}', [ServiceController::class, 'show'])->middleware('auth');
 
-Route::get('/contracts/create/{customer?}', [ContractController::class, 'create'])->name('contracts.create');
-Route::get('/contracts', [ContractController::class, 'index']);
-Route::get('/contracts/{contract}', [ContractController::class, 'show']);
-Route::post('/contracts', [ContractController::class, 'store']);
-Route::patch('/contracts/{contract}', [ContractController::class, 'update']);
+Route::get('/contracts/create/{customer?}', [ContractController::class, 'create'])->name('contracts.create')->middleware('auth');
+Route::get('/contracts', [ContractController::class, 'index'])->middleware('auth');
+Route::get('/contracts/{contract}', [ContractController::class, 'show'])->middleware('auth');
+Route::post('/contracts', [ContractController::class, 'store'])->middleware('auth');
+Route::patch('/contracts/{contract}', [ContractController::class, 'update'])->middleware('auth');
 
-Route::get('/customers', [CustomerController::class, 'index']);
-Route::get('/customers/{customer}', [CustomerController::class, 'show']);
-Route::post('/customers', [CustomerController::class, 'create']);
-Route::patch('/customers/{customer}', [CustomerController::class, 'update']);
+Route::get('/customers', [CustomerController::class, 'index'])->middleware('auth');
+Route::get('/customers/{customer}', [CustomerController::class, 'show'])->middleware('auth');
+Route::post('/customers', [CustomerController::class, 'create'])->middleware('auth');
+Route::patch('/customers/{customer}', [CustomerController::class, 'update'])->middleware('auth');
 
 Route::get('/contacts', function (){
     return view('contacts.index', ['contacts' => Contact::all()]);
-});
+})->middleware('auth');
+
+Route::get('/register', [RegisteredUserController::class, 'create']);
+Route::post('/register', [RegisteredUserController::class, 'store']);
+
+Route::get('/login', [SessionController::class, 'create']);
+Route::post('/login', [SessionController::class, 'store']);
+Route::post('/logout', [SessionController::class, 'destroy']);
 
 ?>
